fix: new join list で unknown を出さない
