+++ electron/module/vrchatPhoto/error.ts
